tests/wpt/tests/html/semantics/permission-element/type-supported-feature-detect.tentative.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Usermedia Element: Feature detection</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  test(() => {
    assert_not_equals(typeof HTMLUserMediaElement, undefined,
        `HTMLUserMediaElement should not be "undefined"`);
    assert_equals(typeof HTMLUserMediaElement.isTypeSupported, "function",
        `The type of the isTypeSupported should be "function"`);
  }, "Test HTMLUserMediaElement and isTypeSupported existence");

  [
    "camera",
    "microphone",
    " camera  microphone ",
    " microphone  camera   ",
    " camera camera microphone",
    "microphone microphone camera "
  ].forEach((type) => {
    test(() => {
      assert_true(HTMLUserMediaElement.isTypeSupported(type));
    }, `Test HTMLUserMediaElement should support type ${type}`);
  });

  [
    "invalid",
    "cameraa",
    "microphone geolocation",
    "camera geolocation",
    "camera geolocation microphone",
  ].forEach((type) => {
    test(() => {
      assert_false(HTMLUserMediaElement.isTypeSupported(type));
    }, `Test HTMLUserMediaElement should not support type ${type}`);
  });
</script>
</body>
